requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
python-multipart==0.0.20
sqlalchemy==2.0.36
chromadb==0.6.3
sentence-transformers==3.3.1
langchain==0.3.14
langchain-community==0.3.14
langchain-text-splitters==0.3.4
openai==1.59.3
youtube-transcript-api==0.6.3
pymupdf==1.25.3
python-pptx==1.0.2
networkx==3.4.2
pyvis==0.3.2
pydantic==2.10.4
scikit-learn==1.6.1
python-dotenv==1.0.1
